`Devise::RegistrationsController` => `PasswordRegistrationsController`
* Following the pattern of prefixing password-specific code in Devise
	with `Password`, this explicitly labels the
	`Devise::RegistrationsController` as
	`Devise::PasswordRegistrationsController`, since it concerns
	registering a user with a password as the authentication method
