+++ app/views/devise/password_registrations/edit.html.erb
