+++ lib/generators/templates/controllers/password_registrations_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class <%= @scope_prefix %>RegistrationsController < Devise::RegistrationsController
+class <%= @scope_prefix %>PasswordRegistrationsController < Devise::PasswordRegistrationsController
   # before_action :configure_sign_up_params, only: [:create]
   # before_action :configure_account_update_params, only: [:update]
