+++ lib/generators/templates/simple_form_for/password_registrations/edit.html.erb
