+++ lib/generators/templates/simple_form_for/password_registrations/new.html.erb
