+++ test/generators/controllers_generator_test.rb
@@ -10,7 +10,7 @@ class ControllersGeneratorTest < Rails::Generators::TestCase
   test "Assert no controllers are created with no params" do
     capture(:stderr) { run_generator }
     assert_no_file "app/controllers/sessions_controller.rb"
-    assert_no_file "app/controllers/registrations_controller.rb"
+    assert_no_file "app/controllers/password_registrations_controller.rb"
     assert_no_file "app/controllers/confirmations_controller.rb"
     assert_no_file "app/controllers/passwords_controller.rb"
     assert_no_file "app/controllers/unlocks_controller.rb"
@@ -28,7 +28,7 @@ class ControllersGeneratorTest < Rails::Generators::TestCase
   test "Assert specified controllers with scope" do
     run_generator %w(users -c sessions)
     assert_file "app/controllers/users/sessions_controller.rb"
-    assert_no_file "app/controllers/users/registrations_controller.rb"
+    assert_no_file "app/controllers/users/password_registrations_controller.rb"
     assert_no_file "app/controllers/users/confirmations_controller.rb"
     assert_no_file "app/controllers/users/passwords_controller.rb"
     assert_no_file "app/controllers/users/unlocks_controller.rb"
@@ -41,7 +41,7 @@ class ControllersGeneratorTest < Rails::Generators::TestCase
       base_dir = "app/controllers#{scope.blank? ? '' : ('/' + scope)}"
       scope_prefix = scope.blank? ? '' : (scope.camelize + '::')
       controllers = options[:controllers] ||
-        %w(confirmations passwords registrations sessions unlocks omniauth_callbacks)
+        %w(confirmations passwords password_registrations sessions unlocks omniauth_callbacks)
 
       controllers.each do |c|
         assert_file "#{base_dir}/#{c}_controller.rb", /#{scope_prefix + c.camelize}/
